access and print student names correctly
fix getStudentsNames() in GradesPanel

this program has to go in order for it to work

+++ src/Assignments.java
@@ -6,7 +6,7 @@ public class Assignments {
     private String assignmentName;
     private double grade;
 
-    public Assignments(String assignmentName, int grade) {
+    public Assignments(String assignmentName, double grade) {
         this.assignmentName = assignmentName;
         this.grade = grade;
     }
